# improve startup performance for clojure apps