--- a/dd-java-agent/agent-tooling/src/main/java/datadog/trace/agent/tooling/bytebuddy/matcher/GlobalIgnoresMatcher.java
+++ b/dd-java-agent/agent-tooling/src/main/java/datadog/trace/agent/tooling/bytebuddy/matcher/GlobalIgnoresMatcher.java
@@ -24,7 +24,7 @@ public class GlobalIgnoresMatcher<T extends TypeDescription>
     extends ElementMatcher.Junction.AbstractBase<T> {
 
   private static final Pattern COM_MCHANGE_PROXY =
-      Pattern.compile("com\\.mchange\\.v2\\.c3p0\\..*Proxy");
+      Pattern.compile("^com\\.mchange\\.v2\\.c3p0\\..*Proxy$");
 
   public static <T extends TypeDescription> ElementMatcher.Junction<T> globalIgnoresMatcher(
       final boolean skipAdditionalLibraryMatcher) {
@@ -47,129 +47,159 @@ public class GlobalIgnoresMatcher<T extends TypeDescription>
   @Override
   public boolean matches(final T target) {
     final String name = target.getActualName();
-
-    if (name.startsWith("datadog.opentracing.")
-        || name.startsWith("datadog.trace.core.")
-        || name.startsWith("datadog.slf4j.")
-        || name.startsWith("net.bytebuddy.")
-        || name.startsWith("jdk.")
-        || name.startsWith("org.aspectj.")
-        || name.startsWith("com.intellij.rt.debugger.")
-        || name.startsWith("com.p6spy.")
-        || name.startsWith("com.newrelic.")
-        || name.startsWith("com.dynatrace.")
-        || name.startsWith("com.jloadtrace.")
-        || name.startsWith("com.appdynamics.")
-        || name.startsWith("com.singularity.")
-        || name.startsWith("com.jinspired.")
-        || name.startsWith("org.jinspired.")) {
-      return true;
+    switch (name.charAt(0) - 'a') {
+        // starting at zero to get a tableswitch from javac, though it looks horrendous
+      case 'a' - 'a':
+        break;
+      case 'b' - 'a':
+        break;
+      case 'c' - 'a':
+        if (name.startsWith("com.")) {
+          if (name.startsWith("com.p6spy.")
+              || name.startsWith("com.newrelic.")
+              || name.startsWith("com.dynatrace.")
+              || name.startsWith("com.jloadtrace.")
+              || name.startsWith("com.appdynamics.")
+              || name.startsWith("com.singularity.")
+              || name.startsWith("com.jinspired.")
+              || name.startsWith("com.intellij.rt.debugger.")) {
+            return true;
+          }
+          if (name.startsWith("com.sun.")) {
+            return !name.startsWith("com.sun.messaging.")
+                && !name.startsWith("com.sun.jersey.api.client");
+          }
+          if (COM_MCHANGE_PROXY.matcher(name).matches()) {
+            return true;
+          }
+        }
+        if (name.startsWith("clojure.")) {
+          return true;
+        }
+        break;
+      case 'd' - 'a':
+        if (name.startsWith("datadog.")) {
+          if (name.startsWith("datadog.opentracing.")
+              || name.startsWith("datadog.trace.core.")
+              || name.startsWith("datadog.slf4j.")) {
+            return true;
+          }
+          if (name.startsWith("datadog.trace.")) {
+            // FIXME: We should remove this once
+            // https://github.com/raphw/byte-buddy/issues/558 is fixed
+            return !name.equals(
+                    "datadog.trace.bootstrap.instrumentation.java.concurrent.RunnableWrapper")
+                && !name.equals(
+                    "datadog.trace.bootstrap.instrumentation.java.concurrent.CallableWrapper");
+          }
+        }
+        break;
+      case 'e' - 'a':
+        break;
+      case 'f' - 'a':
+        break;
+      case 'g' - 'a':
+        break;
+      case 'h' - 'a':
+        break;
+      case 'i' - 'a':
+        break;
+      case 'j' - 'a':
+        if (name.startsWith("jdk.")) {
+          return true;
+        }
+        if (name.startsWith("java.")) {
+          // allow exception profiling instrumentation
+          if (name.equals("java.lang.Throwable")) {
+            return false;
+          }
+          if (name.equals("java.net.URL") || name.equals("java.net.HttpURLConnection")) {
+            return false;
+          }
+          if (name.startsWith("java.rmi.") || name.startsWith("java.util.concurrent.")) {
+            return false;
+          }
+          // Concurrent instrumentation modifies the structure of
+          // Cleaner class incompatibly with java9+ modules.
+          // Working around until a long-term fix for modules can be
+          // put in place.
+          return !name.startsWith("java.util.logging.")
+              || name.equals("java.util.logging.LogManager$Cleaner");
+        }
+        break;
+      case 'k' - 'a':
+        break;
+      case 'l' - 'a':
+        break;
+      case 'm' - 'a':
+        break;
+      case 'n' - 'a':
+        if (name.startsWith("net.bytebuddy.")) {
+          return true;
+        }
+        break;
+      case 'o' - 'a':
+        if (name.startsWith("org.")) {
+          if (name.startsWith("org.aspectj.") || name.startsWith("org.jinspired.")) {
+            return true;
+          }
+          // groovy
+          if (name.startsWith("org.groovy.") || name.startsWith("org.apache.groovy.")) {
+            return true;
+          }
+          if (name.startsWith("org.codehaus.groovy.")) {
+            // We seem to instrument some classes in runtime
+            return !name.startsWith("org.codehaus.groovy.runtime.");
+          }
+          if (name.startsWith("org.apache.log4j.")) {
+            return !name.equals("org.apache.log4j.MDC");
+          }
+        }
+        break;
+      case 'p' - 'a':
+        break;
+      case 'q' - 'a':
+        break;
+      case 'r' - 'a':
+        break;
+      case 's' - 'a':
+        if (name.startsWith("sun.")) {
+          return !name.startsWith("sun.net.www.protocol.")
+              && !name.startsWith("sun.rmi.server")
+              && !name.startsWith("sun.rmi.transport")
+              && !name.equals("sun.net.www.http.HttpClient");
+        }
+        break;
+      default:
     }
 
-    // groovy
-    if (name.startsWith("org.groovy.") || name.startsWith("org.apache.groovy.")) {
-      return true;
-    }
-    if (name.startsWith("org.codehaus.groovy.")) {
-      // We seem to instrument some classes in runtime
-      if (name.startsWith("org.codehaus.groovy.runtime.")) {
-        return false;
+    final int firstDollar = name.indexOf('$');
+    if (firstDollar > -1) {
+      // clojure class patterns
+      if (name.startsWith("loader__", firstDollar + 1)) {
+        return true;
       }
-      return true;
-    }
-    // clojure
-    if (name.startsWith("clojure.") || name.contains("$fn__")) {
-      return true;
-    }
-
-    if (name.startsWith("datadog.trace.")) {
-      // FIXME: We should remove this once
-      // https://github.com/raphw/byte-buddy/issues/558 is fixed
-      if (name.equals("datadog.trace.bootstrap.instrumentation.java.concurrent.RunnableWrapper")
-          || name.equals(
-              "datadog.trace.bootstrap.instrumentation.java.concurrent.CallableWrapper")) {
-        return false;
-      }
-      return true;
-    }
-
-    if (name.startsWith("java.")) {
-      // allow exception profiling instrumentation
-      if (name.equals("java.lang.Throwable")) {
-        return false;
-      }
-      if (name.equals("java.net.URL") || name.equals("java.net.HttpURLConnection")) {
-        return false;
-      }
-      if (name.startsWith("java.rmi.") || name.startsWith("java.util.concurrent.")) {
-        return false;
-      }
-      // Concurrent instrumentation modifies the structure of
-      // Cleaner class incompatibly with java9+ modules.
-      // Working around until a long-term fix for modules can be
-      // put in place.
-      if (name.startsWith("java.util.logging.")
-          && !name.equals("java.util.logging.LogManager$Cleaner")) {
-        return false;
+      int dollar = firstDollar;
+      while (dollar > -1) {
+        if (name.startsWith("fn__", dollar + 1) || name.startsWith("reify__", dollar + 1)) {
+          return true;
+        }
+        dollar = name.indexOf('$', dollar + 1);
       }
 
-      return true;
-    }
-
-    if (name.startsWith("com.sun.")) {
-      if (name.startsWith("com.sun.messaging.") || name.startsWith("com.sun.jersey.api.client")) {
-        return false;
+      if (name.contains("$JaxbAccessor")
+          || name.contains("CGLIB$$")
+          || name.contains("$__sisu")
+          || name.contains("$$EnhancerByProxool$$")
+          || name.startsWith("org.springframework.core.$Proxy")) {
+        return true;
       }
-
+    }
+    if (name.contains("javassist") || name.contains(".asm.")) {
       return true;
     }
 
-    if (name.startsWith("sun.")) {
-      if (name.startsWith("sun.net.www.protocol.")
-          || name.startsWith("sun.rmi.server")
-          || name.startsWith("sun.rmi.transport")
-          || name.equals("sun.net.www.http.HttpClient")) {
-        return false;
-      }
-
-      return true;
-    }
-
-    if (name.startsWith("org.apache.log4j.")) {
-      if (name.equals("org.apache.log4j.MDC")) {
-        return false;
-      }
-
-      return true;
-    }
-
-    if (name.startsWith("org.slf4j.")) {
-      if (name.equals("org.slf4j.MDC")) {
-        return false;
-      }
-
-      return true;
-    }
-
-    if (name.contains("$JaxbAccessor")
-        || name.contains("CGLIB$$")
-        || name.contains("javassist")
-        || name.contains(".asm.")
-        || name.contains("$__sisu")
-        || name.contains("$$EnhancerByProxool$$")
-        || name.startsWith("org.springframework.core.$Proxy")) {
-      return true;
-    }
-
-    if (COM_MCHANGE_PROXY.matcher(name).matches()) {
-      return true;
-    }
-
-    if (!skipAdditionalLibraryMatcher && additionalLibraryIgnoreMatcher.matches(target)) {
-      return true;
-    }
-
-    return false;
+    return !skipAdditionalLibraryMatcher && additionalLibraryIgnoreMatcher.matches(target);
   }
 
   @Override