This currently actually _increases_ the startup time for clojure koans on leiningen by 3 seconds (3s baseline, 13s on master, 16s on this branch, 37s prior to #1642).