@richardstartin 2020-07-01
As of the last push leiningen/clojure koans starts up in under 10s