@richardstartin 2020-07-01
We should rip this out and replace it with a decent trie implementation in the long term, but this is a tactical fix to improve performance in in the clojure applications short term.